producer/requirements.txt
instaloader
kafka-python